pdf_chat/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf==3.17.1
google-generativeai==0.3.0
langchain==0.1.0
pydantic==2.5.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
chromadb==0.4.18 